Split exchanges orchestrator tests in smaller files for better clarity
