+++ CMakeLists.txt
@@ -258,3 +258,4 @@ add_subdirectory(src/objects)
 add_subdirectory(src/api-objects)
 add_subdirectory(src/api)
 add_subdirectory(src/engine)
+add_subdirectory(src/main)
