+++ src/engine/CMakeLists.txt
@@ -1,65 +1,51 @@
 aux_source_directory(src ENGINE_SRC)
 
-if (CCT_BUILD_EXEC)
-  add_executable (coincenter ${ENGINE_SRC})
+add_library(coincenter_engine STATIC ${ENGINE_SRC})
 
-  # Enable LTO with coincenter in Release mode
-  if(CMAKE_BUILD_TYPE STREQUAL "Release")
-    set_property(TARGET coincenter PROPERTY INTERPROCEDURAL_OPTIMIZATION TRUE)
-    message(STATUS "Activate LTO for coincenter")
-  endif()
-
-else()
-
-  list(REMOVE_ITEM ENGINE_SRC "src/main.cpp")
-  add_library(coincenter STATIC ${ENGINE_SRC})
-
-endif()
-
-target_link_libraries(coincenter PUBLIC coincenter_exchangeapi)
-target_link_libraries(coincenter PUBLIC coincenter_objects)
-target_include_directories(coincenter PUBLIC include)
-
-set_target_properties(coincenter PROPERTIES 
-                                 VERSION ${PROJECT_VERSION}
-                                 COMPILE_DEFINITIONS_DEBUG "JSON_DEBUG;JSON_SAFE;JSON_ISO_STRICT"
-                                 RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}
-)    
+target_link_libraries(coincenter_engine PUBLIC coincenter_exchangeapi)
+target_link_libraries(coincenter_engine PUBLIC coincenter_objects)
+target_include_directories(coincenter_engine PUBLIC include)
 
 add_unit_test(
-    commandlineoptionsparser_test
-    test/commandlineoptionsparser_test.cpp
-    DEFINITIONS
-    CCT_DISABLE_SPDLOG
-    LIBRARIES
-    coincenter_tech
+  commandlineoptionsparser_test
+  test/commandlineoptionsparser_test.cpp
+  DEFINITIONS
+  CCT_DISABLE_SPDLOG
+  LIBRARIES
+  coincenter_tech
 )
 
 add_unit_test(
-    exchangesorchestrator_test
-    src/exchangesorchestrator.cpp
-    test/exchangesorchestrator_test.cpp
-    LIBRARIES
-    coincenter_exchangeapi
-    coincenter_objects
+  exchangesorchestrator_private_test
+  test/exchangesorchestrator_private_test.cpp
+  LIBRARIES
+  coincenter_engine
 )
 
 add_unit_test(
-    queryresultprinter_test
-    src/balanceperexchangeportfolio.cpp
-    src/coincentercommandtype.cpp
-    src/queryresultprinter.cpp
-    test/queryresultprinter_test.cpp
-    LIBRARIES
-    coincenter_exchangeapi
-    coincenter_objects
+  exchangesorchestrator_public_test
+  test/exchangesorchestrator_public_test.cpp
+  LIBRARIES
+  coincenter_engine
 )
 
 add_unit_test(
-    stringoptionparser_test
-    src/stringoptionparser.cpp
-    test/stringoptionparser_test.cpp
-    LIBRARIES
-    coincenter_objects
-    coincenter_tools
+  exchangesorchestrator_trade_test
+  test/exchangesorchestrator_trade_test.cpp
+  LIBRARIES
+  coincenter_engine
+)
+
+add_unit_test(
+  queryresultprinter_test
+  test/queryresultprinter_test.cpp
+  LIBRARIES
+  coincenter_engine
+)
+
+add_unit_test(
+  stringoptionparser_test
+  test/stringoptionparser_test.cpp
+  LIBRARIES
+  coincenter_engine
 )
