+++ src/engine/test/exchangesorchestrator_trade_test.cpp
@@ -1,298 +1,18 @@
-#include "exchangesorchestrator.hpp"
-
 #include <gtest/gtest.h>
 
 #include "cct_const.hpp"
 #include "exchangedata_test.hpp"
+#include "exchangesorchestrator.hpp"
 
 namespace cct {
 
-using Deposit = CurrencyExchange::Deposit;
-using Withdraw = CurrencyExchange::Withdraw;
-using Type = CurrencyExchange::Type;
-using TradeInfo = api::TradeInfo;
 using OrderInfo = api::OrderInfo;
 using PlaceOrderInfo = api::PlaceOrderInfo;
-using UniquePublicSelectedExchanges = ExchangeRetriever::UniquePublicSelectedExchanges;
 
 class ExchangeOrchestratorTest : public ExchangesBaseTest {
  protected:
   ExchangesOrchestrator exchangesOrchestrator{std::span<Exchange>(&this->exchange1, 8)};
 };
-
-TEST_F(ExchangeOrchestratorTest, TickerInformation) {
-  const MarketOrderBookMap marketOrderbookMap1 = {{m1, marketOrderBook10}, {m2, marketOrderBook20}};
-  EXPECT_CALL(exchangePublic1, queryAllApproximatedOrderBooks(1)).WillOnce(testing::Return(marketOrderbookMap1));
-
-  const MarketOrderBookMap marketOrderbookMap2 = {{m1, marketOrderBook10}, {m3, marketOrderBook3}};
-  EXPECT_CALL(exchangePublic2, queryAllApproximatedOrderBooks(1)).WillOnce(testing::Return(marketOrderbookMap2));
-
-  const ExchangeName kTestedExchanges12[] = {ExchangeName(kSupportedExchanges[0]),
-                                             ExchangeName(kSupportedExchanges[1])};
-
-  ExchangeTickerMaps expectedTickerMaps = {{&exchange1, marketOrderbookMap1}, {&exchange2, marketOrderbookMap2}};
-  EXPECT_EQ(exchangesOrchestrator.getTickerInformation(kTestedExchanges12), expectedTickerMaps);
-}
-
-class ExchangeOrchestratorMarketOrderbookTest : public ExchangeOrchestratorTest {
- protected:
-  ExchangeOrchestratorMarketOrderbookTest() {
-    EXPECT_CALL(exchangePublic1, queryTradableMarkets()).WillOnce(testing::Return(markets1));
-    EXPECT_CALL(exchangePublic2, queryTradableMarkets()).WillOnce(testing::Return(markets2));
-    EXPECT_CALL(exchangePublic3, queryTradableMarkets()).WillOnce(testing::Return(markets3));
-
-    EXPECT_CALL(exchangePublic1, queryOrderBook(testedMarket, testing::_)).WillOnce(testing::Return(marketOrderBook20));
-    EXPECT_CALL(exchangePublic3, queryOrderBook(testedMarket, testing::_)).WillOnce(testing::Return(marketOrderBook21));
-  }
-
-  Market testedMarket = m2;
-
-  // Mock tradable markets
-  const MarketSet markets1{m1, testedMarket};
-  const MarketSet markets2{m1, m3};
-  const MarketSet markets3{m1, testedMarket, m3};
-  CurrencyCode equiCurrencyCode;
-  std::optional<int> optDepth;
-  MarketOrderBookConversionRates marketOrderBookConversionRates{{exchange1.name(), marketOrderBook20, std::nullopt},
-                                                                {exchange3.name(), marketOrderBook21, std::nullopt}};
-};
-
-TEST_F(ExchangeOrchestratorMarketOrderbookTest, AllSpecifiedExchanges) {
-  const ExchangeName kTestedExchanges123[] = {
-      ExchangeName(kSupportedExchanges[0]), ExchangeName(kSupportedExchanges[1]), ExchangeName(kSupportedExchanges[2])};
-
-  EXPECT_EQ(exchangesOrchestrator.getMarketOrderBooks(testedMarket, kTestedExchanges123, equiCurrencyCode, optDepth),
-            marketOrderBookConversionRates);
-}
-
-TEST_F(ExchangeOrchestratorMarketOrderbookTest, ImplicitAllExchanges) {
-  EXPECT_EQ(exchangesOrchestrator.getMarketOrderBooks(testedMarket, ExchangeNameSpan{}, equiCurrencyCode, optDepth),
-            marketOrderBookConversionRates);
-}
-
-class ExchangeOrchestratorEmptyMarketOrderbookTest : public ExchangeOrchestratorTest {
- protected:
-  ExchangeOrchestratorEmptyMarketOrderbookTest() {
-    EXPECT_CALL(exchangePublic2, queryTradableMarkets()).WillOnce(testing::Return(markets2));
-  }
-
-  Market testedMarket = m2;
-
-  // Mock tradable markets
-  const MarketSet markets1{m1, testedMarket};
-  const MarketSet markets2{m1, m3};
-  const MarketSet markets3{m1, testedMarket, m3};
-  CurrencyCode equiCurrencyCode;
-  std::optional<int> optDepth;
-  MarketOrderBookConversionRates marketOrderBookConversionRates;
-};
-
-TEST_F(ExchangeOrchestratorEmptyMarketOrderbookTest, MarketDoesNotExist) {
-  const ExchangeName kTestedExchanges2[] = {ExchangeName(kSupportedExchanges[1])};
-  EXPECT_EQ(exchangesOrchestrator.getMarketOrderBooks(testedMarket, kTestedExchanges2, equiCurrencyCode, optDepth),
-            marketOrderBookConversionRates);
-}
-
-TEST_F(ExchangeOrchestratorTest, BalanceNoEquivalentCurrencyUniqueExchange) {
-  CurrencyCode equiCurrency;
-
-  EXPECT_CALL(exchangePrivate1, queryAccountBalance(equiCurrency)).WillOnce(testing::Return(balancePortfolio1));
-
-  const ExchangeName privateExchangeNames[1] = {ExchangeName(exchange1.name(), exchange1.keyName())};
-  BalancePerExchange ret{{&exchange1, balancePortfolio1}};
-  EXPECT_EQ(exchangesOrchestrator.getBalance(privateExchangeNames, equiCurrency), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, BalanceNoEquivalentCurrencySeveralExchanges) {
-  CurrencyCode equiCurrency;
-
-  EXPECT_CALL(exchangePrivate1, queryAccountBalance(equiCurrency)).WillOnce(testing::Return(balancePortfolio1));
-  EXPECT_CALL(exchangePrivate3, queryAccountBalance(equiCurrency)).WillOnce(testing::Return(balancePortfolio2));
-  EXPECT_CALL(exchangePrivate4, queryAccountBalance(equiCurrency)).WillOnce(testing::Return(balancePortfolio3));
-
-  const ExchangeName privateExchangeNames[] = {ExchangeName(exchange3.name(), exchange3.keyName()),
-                                               ExchangeName(exchange1.name(), exchange1.keyName()),
-                                               ExchangeName(exchange4.name(), exchange4.keyName())};
-  BalancePerExchange ret{
-      {&exchange1, balancePortfolio1}, {&exchange3, balancePortfolio2}, {&exchange4, balancePortfolio3}};
-  EXPECT_EQ(exchangesOrchestrator.getBalance(privateExchangeNames, equiCurrency), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, DepositInfoUniqueExchanges) {
-  CurrencyCode depositCurrency{"ETH"};
-
-  const ExchangeName privateExchangeNames[] = {ExchangeName(exchange2.name(), exchange2.keyName())};
-
-  CurrencyExchangeFlatSet tradableCurrencies2{CurrencyExchangeVector{
-      CurrencyExchange(depositCurrency, Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange("XRP", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate2, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies2));
-
-  Wallet wallet2{privateExchangeNames[0], depositCurrency, "address1", "", WalletCheck()};
-  EXPECT_CALL(exchangePrivate2, queryDepositWallet(depositCurrency)).WillOnce(testing::Return(wallet2));
-
-  WalletPerExchange ret{{&exchange2, wallet2}};
-  EXPECT_EQ(exchangesOrchestrator.getDepositInfo(privateExchangeNames, depositCurrency), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, DepositInfoSeveralExchangesWithUnavailableDeposits) {
-  CurrencyCode depositCurrency{"XRP"};
-
-  const ExchangeName privateExchangeNames[] = {
-      ExchangeName(exchange3.name(), exchange3.keyName()), ExchangeName(exchange1.name(), exchange1.keyName()),
-      ExchangeName(exchange2.name(), exchange2.keyName()), ExchangeName(exchange4.name(), exchange4.keyName())};
-
-  CurrencyExchangeFlatSet tradableCurrencies1{CurrencyExchangeVector{
-      CurrencyExchange(depositCurrency, Deposit::kUnavailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate1, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies1));
-
-  CurrencyExchangeFlatSet tradableCurrencies2{
-      CurrencyExchangeVector{CurrencyExchange("XLM", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate2, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies2));
-
-  CurrencyExchangeFlatSet tradableCurrencies3{CurrencyExchangeVector{
-      CurrencyExchange("BTC", Deposit::kUnavailable, Withdraw::kUnavailable, Type::kCrypto),
-      CurrencyExchange("SOL", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange(depositCurrency, Deposit::kAvailable, Withdraw::kUnavailable, Type::kCrypto),
-      CurrencyExchange("EUR", Deposit::kAvailable, Withdraw::kAvailable, Type::kFiat),
-  }};
-  EXPECT_CALL(exchangePrivate3, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies3));
-  EXPECT_CALL(exchangePrivate4, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies3));
-
-  Wallet wallet31{privateExchangeNames[2], depositCurrency, "address2", "tag2", WalletCheck()};
-  EXPECT_CALL(exchangePrivate3, queryDepositWallet(depositCurrency)).WillOnce(testing::Return(wallet31));
-
-  Wallet wallet32{privateExchangeNames[3], depositCurrency, "address3", "tag3", WalletCheck()};
-  EXPECT_CALL(exchangePrivate4, queryDepositWallet(depositCurrency)).WillOnce(testing::Return(wallet32));
-
-  WalletPerExchange ret{{&exchange3, wallet31}, {&exchange4, wallet32}};
-  EXPECT_EQ(exchangesOrchestrator.getDepositInfo(privateExchangeNames, depositCurrency), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, GetOpenedOrders) {
-  OrdersConstraints noConstraints;
-
-  const ExchangeName privateExchangeNames[] = {ExchangeName(exchange3.name(), exchange3.keyName()),
-                                               ExchangeName(exchange2.name(), exchange2.keyName()),
-                                               ExchangeName(exchange4.name(), exchange4.keyName())};
-
-  Orders orders2{Order("Id1", MonetaryAmount("0.1ETH"), MonetaryAmount("0.9ETH"), MonetaryAmount("0.14BTC"),
-                       Clock::now(), TradeSide::kBuy),
-                 Order("Id2", MonetaryAmount("15XLM"), MonetaryAmount("76XLM"), MonetaryAmount("0.5EUR"), Clock::now(),
-                       TradeSide::kSell)};
-  EXPECT_CALL(exchangePrivate2, queryOpenedOrders(noConstraints)).WillOnce(testing::Return(orders2));
-
-  Orders orders3{};
-  EXPECT_CALL(exchangePrivate3, queryOpenedOrders(noConstraints)).WillOnce(testing::Return(orders3));
-
-  Orders orders4{Order("Id37", MonetaryAmount("0.7ETH"), MonetaryAmount("0.9ETH"), MonetaryAmount("0.14BTC"),
-                       Clock::now(), TradeSide::kSell),
-                 Order("Id2", MonetaryAmount("15XLM"), MonetaryAmount("19XLM"), MonetaryAmount("0.5EUR"), Clock::now(),
-                       TradeSide::kBuy)};
-  EXPECT_CALL(exchangePrivate4, queryOpenedOrders(noConstraints)).WillOnce(testing::Return(orders4));
-
-  OpenedOrdersPerExchange ret{{&exchange2, OrdersSet(orders2.begin(), orders2.end())},
-                              {&exchange3, OrdersSet(orders3.begin(), orders3.end())},
-                              {&exchange4, OrdersSet(orders4.begin(), orders4.end())}};
-  EXPECT_EQ(exchangesOrchestrator.getOpenedOrders(privateExchangeNames, noConstraints), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, GetMarketsPerExchangeOneCurrency) {
-  CurrencyCode cur1{"LUNA"};
-  CurrencyCode cur2{};
-  ExchangeNameSpan exchangeNameSpan{};
-
-  Market m4{"LUNA", "BTC"};
-  Market m5{"SHIB", "LUNA"};
-  Market m6{"DOGE", "EUR"};
-
-  MarketSet markets1{m1, m2, m4, m6};
-  EXPECT_CALL(exchangePublic1, queryTradableMarkets()).WillOnce(testing::Return(markets1));
-  MarketSet markets2{m1, m2, m3, m4, m5, m6};
-  EXPECT_CALL(exchangePublic2, queryTradableMarkets()).WillOnce(testing::Return(markets2));
-  MarketSet markets3{m1, m2, m6};
-  EXPECT_CALL(exchangePublic3, queryTradableMarkets()).WillOnce(testing::Return(markets3));
-
-  MarketsPerExchange ret{{&exchange1, MarketSet{m4}}, {&exchange2, MarketSet{m4, m5}}, {&exchange3, MarketSet{}}};
-  EXPECT_EQ(exchangesOrchestrator.getMarketsPerExchange(cur1, cur2, exchangeNameSpan), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, GetMarketsPerExchangeTwoCurrencies) {
-  CurrencyCode cur1{"LUNA"};
-  CurrencyCode cur2{"SHIB"};
-  ExchangeNameSpan exchangeNameSpan{};
-
-  Market m4{"LUNA", "BTC"};
-  Market m5{"SHIB", "LUNA"};
-  Market m6{"DOGE", "EUR"};
-  Market m7{"LUNA", "EUR"};
-
-  MarketSet markets1{m1, m2, m4, m6, m7};
-  EXPECT_CALL(exchangePublic1, queryTradableMarkets()).WillOnce(testing::Return(markets1));
-  MarketSet markets2{m1, m2, m3, m4, m5, m6};
-  EXPECT_CALL(exchangePublic2, queryTradableMarkets()).WillOnce(testing::Return(markets2));
-  MarketSet markets3{m1, m2, m6, m7};
-  EXPECT_CALL(exchangePublic3, queryTradableMarkets()).WillOnce(testing::Return(markets3));
-
-  MarketsPerExchange ret{{&exchange1, MarketSet{}}, {&exchange2, MarketSet{m5}}, {&exchange3, MarketSet{}}};
-  EXPECT_EQ(exchangesOrchestrator.getMarketsPerExchange(cur1, cur2, exchangeNameSpan), ret);
-}
-
-TEST_F(ExchangeOrchestratorTest, GetExchangesTradingCurrency) {
-  CurrencyCode currencyCode{"XRP"};
-
-  const ExchangeName kTestedExchanges13[] = {ExchangeName(kSupportedExchanges[0]),
-                                             ExchangeName(kSupportedExchanges[2])};
-
-  CurrencyExchangeFlatSet tradableCurrencies1{
-      CurrencyExchangeVector{CurrencyExchange("XRP", Deposit::kUnavailable, Withdraw::kAvailable, Type::kCrypto),
-                             CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate1, queryTradableCurrencies())
-      .Times(2)
-      .WillRepeatedly(testing::Return(tradableCurrencies1));
-
-  CurrencyExchangeFlatSet tradableCurrencies3{CurrencyExchangeVector{
-      CurrencyExchange("BTC", Deposit::kUnavailable, Withdraw::kUnavailable, Type::kCrypto),
-      CurrencyExchange("SOL", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange("XRP", Deposit::kAvailable, Withdraw::kUnavailable, Type::kCrypto),
-      CurrencyExchange("EUR", Deposit::kAvailable, Withdraw::kAvailable, Type::kFiat),
-  }};
-  EXPECT_CALL(exchangePrivate3, queryTradableCurrencies())
-      .Times(2)
-      .WillRepeatedly(testing::Return(tradableCurrencies3));
-
-  UniquePublicSelectedExchanges ret1{&exchange1, &exchange3};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingCurrency(currencyCode, kTestedExchanges13, false), ret1);
-  UniquePublicSelectedExchanges ret2{&exchange1};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingCurrency(currencyCode, kTestedExchanges13, true), ret2);
-}
-
-TEST_F(ExchangeOrchestratorTest, GetExchangesTradingMarket) {
-  constexpr int kNbTests = 5;
-  const MarketSet markets1{Market("SOL", "BTC"),   Market("XRP", "BTC"),   Market("XRP", "EUR"),
-                           Market("SHIB", "DOGE"), Market("SHIB", "USDT"), Market("XLM", "BTC")};
-  EXPECT_CALL(exchangePublic1, queryTradableMarkets()).Times(kNbTests).WillRepeatedly(testing::Return(markets1));
-
-  const MarketSet markets2{Market("SOL", "BTC"), Market("XRP", "BTC"), Market("XRP", "KRW"), Market("SHIB", "KRW"),
-                           Market("XLM", "KRW")};
-  EXPECT_CALL(exchangePublic2, queryTradableMarkets()).Times(kNbTests).WillRepeatedly(testing::Return(markets2));
-
-  const MarketSet markets3{Market("LUNA", "BTC"), Market("AVAX", "USD"), Market("SOL", "BTC"), Market("XRP", "BTC"),
-                           Market("XRP", "KRW"),  Market("SHIB", "KRW"), Market("XLM", "BTC")};
-  EXPECT_CALL(exchangePublic3, queryTradableMarkets()).Times(kNbTests).WillRepeatedly(testing::Return(markets3));
-
-  UniquePublicSelectedExchanges ret1{&exchange1, &exchange2, &exchange3};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingMarket(Market("SOL", "BTC"), ExchangeNameSpan{}), ret1);
-  UniquePublicSelectedExchanges ret2{&exchange3};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingMarket(Market("AVAX", "USD"), ExchangeNameSpan{}), ret2);
-  UniquePublicSelectedExchanges ret3{};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingMarket(Market("SHIB", "EUR"), ExchangeNameSpan{}), ret3);
-  UniquePublicSelectedExchanges ret4{};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingMarket(Market("BTC", "SOL"), ExchangeNameSpan{}), ret4);
-  UniquePublicSelectedExchanges ret5{&exchange1, &exchange3};
-  EXPECT_EQ(exchangesOrchestrator.getExchangesTradingMarket(Market("XLM", "BTC"), ExchangeNameSpan{}), ret5);
-}
 
 /// For the trade tests, 'exchangeprivateapi_test' already tests a lot of complex trade options.
 /// Here we are only interested in testing the orchestrator, so we for simplicity reasons we will do only taker trades.
@@ -1144,126 +864,5 @@ TEST_F(ExchangeOrchestratorTradeTest, SmartSellAllExchanges) {
   TradedAmountsPerExchange tradedAmountsPerExchange{std::make_pair(&exchange1, tradedAmounts1)};
   EXPECT_EQ(tradedAmountsPerExchange,
             exchangesOrchestrator.smartSell(startAmount, false, ExchangeNames{}, tradeOptions));
-}
-
-TEST_F(ExchangeOrchestratorTest, WithdrawSameAccountImpossible) {
-  MonetaryAmount grossAmount{1000, "XRP"};
-  ExchangeName fromExchange(exchange1.name(), exchange1.keyName());
-  const ExchangeName &toExchange = fromExchange;
-  EXPECT_THROW(exchangesOrchestrator.withdraw(grossAmount, false, fromExchange, toExchange), exception);
-}
-
-TEST_F(ExchangeOrchestratorTest, WithdrawImpossibleFrom) {
-  MonetaryAmount grossAmount{1000, "XRP"};
-  ExchangeName fromExchange(exchange1.name(), exchange1.keyName());
-  ExchangeName toExchange(exchange2.name(), exchange2.keyName());
-
-  CurrencyExchangeFlatSet tradableCurrencies1{CurrencyExchangeVector{
-      CurrencyExchange(grossAmount.currencyCode(), Deposit::kAvailable, Withdraw::kUnavailable, Type::kCrypto),
-      CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate1, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies1));
-  CurrencyExchangeFlatSet tradableCurrencies2{CurrencyExchangeVector{
-      CurrencyExchange(grossAmount.currencyCode(), Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate2, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies2));
-
-  EXPECT_FALSE(exchangesOrchestrator.withdraw(grossAmount, false, fromExchange, toExchange).hasBeenInitiated());
-}
-
-TEST_F(ExchangeOrchestratorTest, WithdrawImpossibleTo) {
-  MonetaryAmount grossAmount{1000, "XRP"};
-  ExchangeName fromExchange(exchange1.name(), exchange1.keyName());
-  ExchangeName toExchange(exchange2.name(), exchange2.keyName());
-
-  CurrencyExchangeFlatSet tradableCurrencies1{CurrencyExchangeVector{
-      CurrencyExchange(grossAmount.currencyCode(), Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate1, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies1));
-  CurrencyExchangeFlatSet tradableCurrencies2{CurrencyExchangeVector{
-      CurrencyExchange(grossAmount.currencyCode(), Deposit::kUnavailable, Withdraw::kAvailable, Type::kCrypto),
-      CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-  EXPECT_CALL(exchangePrivate2, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies2));
-
-  EXPECT_FALSE(exchangesOrchestrator.withdraw(grossAmount, false, fromExchange, toExchange).hasBeenInitiated());
-}
-
-inline bool operator==(const WithdrawInfo &lhs, const WithdrawInfo &rhs) {
-  return lhs.withdrawId() == rhs.withdrawId();
-}
-
-namespace api {
-inline bool operator==(const InitiatedWithdrawInfo &lhs, const InitiatedWithdrawInfo &rhs) {
-  return lhs.withdrawId() == rhs.withdrawId();
-}
-
-inline bool operator==(const SentWithdrawInfo &lhs, const SentWithdrawInfo &rhs) {
-  return lhs.isWithdrawSent() == rhs.isWithdrawSent() && lhs.netEmittedAmount() == rhs.netEmittedAmount();
-}
-}  // namespace api
-
-class ExchangeOrchestratorWithdrawTest : public ExchangeOrchestratorTest {
- protected:
-  ExchangeOrchestratorWithdrawTest() {
-    CurrencyExchangeFlatSet tradableCurrencies1{
-        CurrencyExchangeVector{CurrencyExchange(cur, Deposit::kUnavailable, Withdraw::kAvailable, Type::kCrypto),
-                               CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-    CurrencyExchangeFlatSet tradableCurrencies2{
-        CurrencyExchangeVector{CurrencyExchange(cur, Deposit::kAvailable, Withdraw::kUnavailable, Type::kCrypto),
-                               CurrencyExchange("SHIB", Deposit::kAvailable, Withdraw::kAvailable, Type::kCrypto)}};
-
-    EXPECT_CALL(exchangePrivate1, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies1));
-    EXPECT_CALL(exchangePrivate2, queryTradableCurrencies()).WillOnce(testing::Return(tradableCurrencies2));
-  }
-
-  WithdrawInfo createWithdrawInfo(MonetaryAmount grossAmount, bool isPercentageWithdraw) {
-    if (isPercentageWithdraw) {
-      EXPECT_CALL(exchangePrivate1, queryAccountBalance(CurrencyCode())).WillOnce(testing::Return(balancePortfolio1));
-      grossAmount = (grossAmount.toNeutral() * balancePortfolio1.get(cur)) / 100;
-    } else {
-      EXPECT_CALL(exchangePrivate1, queryAccountBalance(testing::_)).Times(0);
-    }
-    MonetaryAmount netEmittedAmount = grossAmount - fee;
-    Wallet receivingWallet{toExchange, cur, address, tag, WalletCheck()};
-    EXPECT_CALL(exchangePrivate2, queryDepositWallet(cur)).WillOnce(testing::Return(receivingWallet));
-
-    api::InitiatedWithdrawInfo initiatedWithdrawInfo{receivingWallet, withdrawIdView, grossAmount};
-    EXPECT_CALL(exchangePrivate1, launchWithdraw(grossAmount, std::move(receivingWallet)))
-        .WillOnce(testing::Return(initiatedWithdrawInfo));
-    api::SentWithdrawInfo sentWithdrawInfo{netEmittedAmount, true};
-    EXPECT_CALL(exchangePrivate1, isWithdrawSuccessfullySent(initiatedWithdrawInfo))
-        .WillOnce(testing::Return(sentWithdrawInfo));
-    EXPECT_CALL(exchangePrivate2, isWithdrawReceived(initiatedWithdrawInfo, sentWithdrawInfo))
-        .WillOnce(testing::Return(true));
-    return WithdrawInfo(initiatedWithdrawInfo, sentWithdrawInfo);
-  }
-
-  CurrencyCode cur{"XRP"};
-  ExchangeName fromExchange{exchange1.name(), exchange1.keyName()};
-  ExchangeName toExchange{exchange2.name(), exchange2.keyName()};
-
-  std::string_view address{"TestAddress"};
-  std::string_view tag{"TestTag"};
-
-  WithdrawIdView withdrawIdView{"WithdrawId"};
-
-  MonetaryAmount fee{"0.02", cur};
-};
-
-TEST_F(ExchangeOrchestratorWithdrawTest, WithdrawPossible) {
-  MonetaryAmount grossAmount{1000, cur};
-  bool isPercentageWithdraw = false;
-  auto exp = createWithdrawInfo(grossAmount, isPercentageWithdraw);
-  auto ret =
-      exchangesOrchestrator.withdraw(grossAmount, isPercentageWithdraw, fromExchange, toExchange, Duration::zero());
-  EXPECT_EQ(exp, ret);
-}
-
-TEST_F(ExchangeOrchestratorWithdrawTest, WithdrawPossiblePercentage) {
-  MonetaryAmount grossAmount{25, cur};
-  bool isPercentageWithdraw = true;
-  auto exp = createWithdrawInfo(grossAmount, isPercentageWithdraw);
-  auto ret =
-      exchangesOrchestrator.withdraw(grossAmount, isPercentageWithdraw, fromExchange, toExchange, Duration::zero());
-  EXPECT_EQ(exp, ret);
 }
 }  // namespace cct
