+++ src/main/src/processcommandsfromcli.cpp
